Made image ids unsigned, added item limit to storage

+++ common/logic/include/character/Character.hpp
@@ -22,7 +22,7 @@ class Character : public GameEntity, public Resistible, public StatBased {
 public:
     Character() = default;
 
-    Character(size_t id, std::string_view name, int image_id, unsigned int max_hp,
+    Character(size_t id, std::string_view name, size_t image_id, unsigned int max_hp,
               int exp = kDefaultExp, unsigned int ap = kDefaultActionPoints,
               unsigned int ac = kDefaultArmorClass, const Stats& stats = Stats{},
               const Info& info = {});
