+++ common/logic/include/character/NPC.hpp
@@ -12,7 +12,7 @@ class NPC : public Character {
  public:
     NPC() = default;
 
-    NPC(size_t id, std::string_view name, int image_id, unsigned int max_hp,
+    NPC(size_t id, std::string_view name, size_t image_id, unsigned int max_hp,
         int exp = kDefaultExp, unsigned int ap = kDefaultActionPoints,
         unsigned int ac = kDefaultArmorClass, const Stats& stats = Stats{},
         const SharedStorage<Skill>& skills = {}, const Info& info = {}) :
