+++ common/logic/include/items/Spell.hpp
@@ -10,7 +10,7 @@ class Spell : public GameEntity, public Activatable {
  public:
     Spell() = default;
 
-    Spell(size_t id, std::string_view name, int image_id,
+    Spell(size_t id, std::string_view name, size_t image_id,
           const std::vector<Action>& actions, unsigned int action_cost,
           std::string_view scaling, unsigned int spell_cost,
           Cast cast_type = Cast::Tile, const Info& info = {}) :
@@ -18,7 +18,7 @@ class Spell : public GameEntity, public Activatable {
         Activatable(actions, action_cost, scaling, cast_type),
         spell_cost_(spell_cost) {}
 
-    Spell(size_t id, std::string_view name, int image_id,
+    Spell(size_t id, std::string_view name, size_t image_id,
           std::vector<Action>&& actions, unsigned int action_cost,
           std::string_view scaling, unsigned int spell_cost,
           Cast cast_type = Cast::Tile, const Info& info = {}) :
