+++ common/logic/include/structure/Location.hpp
@@ -52,7 +52,7 @@ class Location : public GameEntity {
  public:
     Location() = default;
 
-    Location(size_t id, std::string_view name, int image_id,
+    Location(size_t id, std::string_view name, size_t image_id,
              size_t height, size_t width, const Info& info = {});
 
     ErrorStatus setHeight(size_t height);
